# [WB-5641 and WB-6016] Reduce Memory Footprint of Images In Tables