@tssweeney 2021-07-15
> Looks good. The only cavaet is if someone calls `to_data_array` all the images will be loaded back into memory but if this is rare, that should be ok.

Fair enough (: - I added to to the end of that function as well.